requirements.txt
streamlit
google-genai
imapclient
python-dotenv
